order chưa tích hợp thanh toán stripe

+++ Online-food-ordering-service/src/main/java/com/foodgo/model/Order.java
@@ -38,7 +38,7 @@ public class Order {
     private List<OrderItem> items;
 
     //private Payment payment;
-    private int totalItem;
+    private Long totalItem;
 
     private Long totalPrice;
 }
